SmartQuiz-main/requirements.txt
Flask==3.0.3
pymongo==4.7.3
gunicorn==22.0.0
python-dotenv==1.0.1
